fix(screen): generate kiln canvas in container
BREAKING CHANGE: you will now need to specify a
div element rather than a canvas element when
initializing KGames. The passed div will be filled
with a canvas element respecting the height and
width of the container class.

this allows for us to appropriately resize canvas
and properly track mouse position regardless of
canvas size. This is good because we can have both
fixed and variable width games now.

Fixes #103 #12